parisc: dino_pci: Use pci_host_resource_survey()
Now that code is a bit obscure, so I hope I didn't break anything.

It tries to allocate resources from pcibios_fixup_bus(). I ripped that
out, in favor of a call to pci_host_resource_survey() with the
policy set to reassign everything, which is I think what the old
code was attempting to do, except the old code was preserving
the existing p2p bridge sizes, while this will recalculate and
reassign them completely.

Testing needed...

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>