PCI: Define resource allocation policy in the host bridge
This replaces the simple "preserve_config" in struct pci_host_bridge
with a more complete enumeration representing the 4 PCI resource
allocation policies I've identified accross the various architectures
and platforms.

The default is set to "claim and reassign" which corresponds to the
standard x86 behaviour unless global PCI flags such as PCI_PROBE_ONLY
and PCI_REASSIGN_ALL_RSRC are set in which case it is adjusted
accordingly.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>